Increase number of max. visible recipe
Currently, only 20 recipes are displayed. As we do not have more
details specifications we should increase the number a little to
mae sure enough recipes are displayed. In the future pagination
would make the most sense. However, that is outside the scope of
the PR.

This changes the displayed recipes list from 20 to 100.